split up data in preparation for more maps
