+++ src/components/ui/search/SearchResultsList.svelte
@@ -1,6 +1,6 @@
 <script>
   import SearchResult from "./SearchResult.svelte";
-  import { keys } from "$data/keys";
+  import { keys } from "$data/al-mazrah/locations/keys";
   import { searchTerm, selectedMarker, favorites, filter } from "$store";
 
   let filteredKeys = keys;
