+++ src/data/al-mazrah/locations/pois.js
