+++ src/data/maps.js
@@ -1,0 +1,8 @@
+import { data as alMazrahData } from "./al-mazrah/data";
+
+export const maps = {
+  "al-mazrah": {
+    title: "Al Mazrah",
+    data: alMazrahData,
+  },
+};
